pyproject.toml
[project]
name = "openai-agents-examples"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=6.0.2",
    "openai-agents>=0.1.0",
    "python-dotenv>=1.2.1",
]
